fix: address review findings round 2 (#1332)
