+++ pyproject.toml
@@ -16,7 +16,7 @@ dependencies = [
     "google-auth-httplib2>=0.3,<1",
     "google-genai>=1,<2",
     "openai>=2,<3",
-    "Pillow>=12,<13",
+    "Pillow>=12.2.0,<13",
     "python-dotenv>=1,<2",
     "pandas>=3,<4",
     "matplotlib>=3,<4",
